chore: move reducer/actions/types
